+++ src/base_types.jl
@@ -15,3 +15,11 @@ Base.broadcastable(gen::AbstractGenerator) = Ref(gen)
 
 "An abstract type that serves as the base type for convergence objects."
 abstract type AbstractConvergence end
+
+"An abstract type for penalty functions."
+abstract type AbstractPenalty end
+
+"Treat `AbstractPenalty` as scalar when broadcasting."
+Base.broadcastable(pen::AbstractPenalty) = Ref(pen)
+
+const PenaltyOrFun = Union{Function,AbstractPenalty}