+++ src/counterfactuals/generate_counterfactual.jl
@@ -97,7 +97,7 @@ function generate_counterfactual(
     )
 
     # Check for redundancy (assess if already converged with respect to factual):
-    if Convergence.converged(ce.convergence, ce, ce.x)
+    if Convergence.converged(ce.convergence, ce, ce.factual)
         @info "Factual already in target class and probability exceeds threshold γ=$(ce.convergence.decision_threshold)."
         return ce
     end