+++ src/counterfactuals/initialisation.jl
@@ -11,26 +11,28 @@ Initializes the starting point for the factual(s):
 function initialize_state(ce::CounterfactualExplanation)
     @assert ce.initialization ∈ [:identity, :add_perturbation]
 
-    s′ = ce.s′
+    counterfactual_state = ce.counterfactual_state
 
     # No perturbation:
     if ce.initialization == :identity
-        return s′
+        return counterfactual_state
     end
 
     # If latent space, initial point is random anyway:
     if ce.generator.latent_space
-        return s′
+        return counterfactual_state
     end
 
     # Add random perturbation following Slack (2021): https://arxiv.org/abs/2106.02666
     if ce.initialization == :add_perturbation
-        Δs′ = randn(eltype(s′), size(s′)) * convert(eltype(s′), 0.1)
-        Δs′ = apply_mutability(ce, Δs′)
-        s′ .+= Δs′
+        Δcounterfactual_state =
+            randn(eltype(counterfactual_state), size(counterfactual_state)) *
+            convert(eltype(counterfactual_state), 0.1)
+        Δcounterfactual_state = apply_mutability(ce, Δcounterfactual_state)
+        counterfactual_state .+= Δcounterfactual_state
     end
 
-    return s′
+    return counterfactual_state
 end
 
 """
@@ -39,7 +41,7 @@ end
 Initializes the starting point for the factual(s) in-place.
 """
 function initialize_state!(ce::CounterfactualExplanation)
-    ce.s′ = initialize_state(ce)
+    ce.counterfactual_state = initialize_state(ce)
 
     return ce
 end