+++ src/objectives/penalties.jl
@@ -1,4 +1,6 @@
 using CounterfactualExplanations: polynomial_decay
+using CounterfactualExplanations.Models
+using EnergySamplers: EnergySamplers
 using LinearAlgebra: LinearAlgebra, det, norm
 using Random: Random
 using Statistics: mean
@@ -73,7 +75,7 @@ Evaluates how diverse the counterfactuals are using a Determinantal Point Proces
 function ddp_diversity(
     ce::AbstractCounterfactualExplanation; perturbation_size=1e-3, agg=det
 )
-    X = ce.s′
+    X = ce.counterfactual_state
     xs = eachslice(X; dims=ndims(X))
     K = [1 / (1 + LinearAlgebra.norm(x .- y)) for x in xs, y in xs]
     K += LinearAlgebra.Diagonal(
@@ -208,9 +210,9 @@ function energy_constraint(
 
     # Setup:
     ℒ = 0
-    x′ = CounterfactualExplanations.decode_state(ce)     # current state
+    cf = CounterfactualExplanations.decode_state(ce)     # current state
     t = get_target_index(ce.data.y_levels, ce.target)
-    xs = eachslice(x′; dims=ndims(x′))
+    xs = eachslice(cf; dims=ndims(cf))
 
     # Multiplier ϕ for the energy constraint:
     max_steps = CounterfactualExplanations.Convergence.max_iter(ce.convergence)
@@ -232,4 +234,43 @@ function energy_constraint(
     end
 
     return ℒ
+end
+
+struct EnergyDifferential <: AbstractPenalty
+    K::Int
+    agg::Function
+end
+
+EnergyDifferential(; K::Int=50, agg::Function=mean) = EnergyDifferential(K, agg)
+
+function (pen::EnergyDifferential)(ce::AbstractCounterfactualExplanation)
+
+    # If the potential neighbours have not been computed, do so:
+    get!(
+        ce.search,
+        :potential_neighbours,
+        CounterfactualExplanations.find_potential_neighbours(ce, pen.K),
+    )
+
+    # Get potential neighbours:
+    ys = ce.search[:potential_neighbours]
+    if pen.K > size(ys, 2)
+        @warn "`K` is larger than the number of potential neighbours. Future warnings will be suppressed." maxlog =
+            1
+    end
+
+    # Get counterfactual:
+    cf = CounterfactualExplanations.decode_state(ce)     # current state
+    xs = eachslice(cf; dims=ndims(cf))
+
+    # Compute energy differential:
+    Δ = pen.agg(EnergySamplers.energy_differential.(ce.M, xs, (ys,), ce.target))
+
+    return Δ
+end
+
+function EnergySamplers.energy_differential(M::AbstractModel, xgen, xsampled, y::Int)
+    typeof(M.type) <: Models.AbstractFluxNN || throw(NotImplementedModel(M))
+    f = M.fitresult.fitresult
+    return EnergySamplers.energy_differential(f, xgen, xsampled, y)
 end